+++ NEWS.md
@@ -1,3 +1,7 @@
+# socialranking 0.1.0-9001 (change this part until 0.1.1 is released)
+
+* Added parameter `equivalenceClasses` to `newPowerRelation`
+
 # socialranking 0.1.0 (2022-03-28)
 
 * Initial release